+++ backend/src/main/java/pl/factoryofthefuture/factorymanagement/service/EmployeeService.java
@@ -4,10 +4,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pl.factoryofthefuture.factorymanagement.entity.Employee;
+import pl.factoryofthefuture.factorymanagement.exception.NotFoundException;
 import pl.factoryofthefuture.factorymanagement.repository.EmployeeRepository;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 @RequiredArgsConstructor
 @Service
@@ -21,7 +21,7 @@ public class EmployeeService {
 
     public Employee getEmployee(Long id) {
         return employeeRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("No such element " + id));
+                .orElseThrow(() -> new NotFoundException(id));
     }
 
     public Employee saveEmployee(Employee employee) {
@@ -31,7 +31,7 @@ public class EmployeeService {
     @Transactional
     public Employee updateEmployee(Employee employee) {
         Employee updatedEmployee = employeeRepository.findById(employee.getId())
-                .orElseThrow(() -> new NoSuchElementException("Employee not found with id: " + employee.getId()));
+                .orElseThrow(() -> new NotFoundException(employee.getId()));
         updatedEmployee.setName(employee.getName());
         updatedEmployee.setSurname(employee.getSurname());
         updatedEmployee.setJobPosition(employee.getJobPosition());
@@ -46,7 +46,11 @@ public class EmployeeService {
         return employeeRepository.save(updatedEmployee);
     }
 
+    @Transactional
     public void deleteById(long id) {
+        if (!employeeRepository.existsById(id)) {
+            throw new NotFoundException(id);
+        }
         employeeRepository.deleteById(id);
     }
 }
